requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-Bcrypt
Flask-Mail
Flask-WTF
pyodbc
waitress
pandas
WeasyPrint
zeep
requests
requests_pkcs12
python-dotenv
lxml
brazilfiscalreport
pytz
urllib3